move InferenceRunner to a separate file
